+++ inst/gui/gui.r
@@ -1,320 +1,481 @@
 # 3D-Visualization of Geoelectric Resistivity Measurement Profiles
 # Graphical User Interface
-# Anja Kleebaum
+# Author: Anja Kleebaum
 
 ### load packages
 library(geoelectrics)
 require(tcltk)
 library(tkrplot)
 
-### looking for working directory
-whereFrom=as.character(sys.calls()[[1]][2]) 
+### looking for working directory and change working directory to location of this file
+try(setwd(dirname(as.character(sys.calls()[[1]][2]))))
 
-### change working directory to location of this file
-try(setwd(dirname(whereFrom)))
+topLevelWindow <- tktoplevel()
+tkwm.minsize(topLevelWindow, 400, 300)
+tktitle(topLevelWindow) <-
+  '3D-Visualization of Geoelectric Resistivity Measurement Profiles'
 
-tt <- tktoplevel()
-tkwm.minsize(tt, 400, 300)
-tktitle(tt) <- "3D-Visualization of Geoelectric Resistivity Measurement Profiles"
+### Load example when starting
+data(sinkhole)
+profiles <-
+  list(sinkhole@profiles[[1]], sinkhole@profiles[[2]], sinkhole@profiles[[3]])
+selectedProfileNumber <- tclVar(1)
 
 ###---Plotting functions in GUI---###
-plotXyzGui <- function(Profile) {
+plotProcessedDataGui <- function(profile) {
   try(tkgrid.remove(img))
-  img <<- tkrplot(frameRight, function() { plotXyz(Profile) })
+  img <- tkrplot(frameRight, function() {
+    plot(
+      profile,
+      dataType = 'processed',
+      withTopo = FALSE,
+      main = profile@title,
+      ylab = 'Depth [m]'
+    )
+  })
+  assign("img", img, envir = .GlobalEnv)
   tkgrid(img)
 }
 
-levelplotXyzGui <- function(Profile) {
-  img <<- levelplotXyz(Profile)
+levelplotProcessedDataGui <- function(profile) {
+  img <- levelplot(
+    profile,
+    dataType = 'processed',
+    withTopo = FALSE,
+    main = profile@title
+  )
+  assign("img", img, envir = .GlobalEnv)
+  print(img)
 }
 
-plotXyzHeightGui <- function(Profile) {
+plotProcessedDataWithTopoGui <- function(profile) {
   try(tkgrid.remove(img))
-  img <<- tkrplot(frameRight, function() { plotXyzHeight(Profile) })
+  img <- tkrplot(frameRight, function() {
+    plot(profile, dataType = 'processed', withTopo = TRUE)
+  })
+  assign("img", img, envir = .GlobalEnv)
   tkgrid(img)
 }
 
-levelplotXyzHeightGui <- function(Profile) {
-  img <<- levelplotXyzHeight(Profile)
-  #source("savePlots.r", echo=T)
+levelplotProcessedDataWithTopoGui <- function(profile) {
+  img <- levelplot(profile, dataType = 'processed', withTopo = TRUE)
+  assign("img", img, envir = .GlobalEnv)
+  print(img)
 }
 
-levelplotRawGui <- function(Profile) {
-  img <<- levelplotRaw(Profile)
-  #source("savePlots.r", echo=T)
+levelplotRawDataGui <- function(profile) {
+  img <- levelplot(profile, dataType = 'raw', withTopo = FALSE)
+  assign("img", img, envir = .GlobalEnv)
+  print(img)
 }
 
-plotRawGui <- function(Profile) {  
+plotRawDataGui <- function(profile) {
   try(tkgrid.remove(img))
-  img <<- tkrplot(frameRight, function() { plotRaw(Profile) })  
-  tkgrid(img)  
+  img <- tkrplot(frameRight, function() {
+    plot(
+      profile,
+      dataType = 'raw',
+      withTopo = FALSE,
+      main = profile@title,
+      ylab = 'Depth [m]'
+    )
+  })
+  assign("img", img, envir = .GlobalEnv)
+  tkgrid(img)
 }
 
-choosePlottingProfile <- function(type) { 
-  resetFrames()  
-  comboMenu <- comboMenuProfiles(tt)
-  tkgrid(tklabel(tt, text="Profil Nummer: "), comboMenu)
-  tkbind(comboMenu, "<<ComboboxSelected>>", function() type(p[[as.integer(tclvalue(number))]]))
-}
-
-comboMenuProfiles <- function(widget) {
-  items <- c() 
-  for(x in p) {
-    items <- c(items, x@number)
-  }    
-  number <<- tclVar()
-  return(ttkcombobox(widget, values=items, textvariable=number, width=3))  
-}
-
-heightAdjustmentGui <- function() {
-  h <- tktoplevel(tt)
-  tktitle(h) <- "Height Adjustment"
+createComboBoxForProfiles <- function(widget, onChangedSelection) {
+  items <- c()
+  for (profile in profiles) {
+    items <- c(items, profile@number)
+  }
+  selectedProfileNumber <<- tclVar(1)
   
-  comboMenu <- comboMenuProfiles(h)
+  try(tkdestroy(comboBox))
+  comboBox <- ttkcombobox(widget,
+                          values = items,
+                          textvariable = selectedProfileNumber,
+                          width = 3)
+  assign('comboBox', comboBox, envir = .GlobalEnv)
+  tkbind(comboBox, '<<ComboboxSelected>>', function() {
+    onChangedSelection(as.integer(tclvalue(selectedProfileNumber)))
+  })
+  
+  try(tkdestroy(comboBoxLabel))
+  comboBoxLabel <- tklabel(widget, text = 'Profile Number: ')
+  assign('comboBoxLabel', comboBoxLabel, envir = .GlobalEnv)
+  
+  tkgrid(comboBoxLabel, sticky = 'w')
+  tkgrid(comboBox, sticky = 'w', pady = c(0, 10))
+  
+  onChangedSelection(1)
+}
+
+adjustHeightGui <- function() {
+  adjustHeightDialog <- tktoplevel(topLevelWindow)
+  tktitle(adjustHeightDialog) <- 'Height Adjustment'
+  
+  onChangedSelection <- function(selectedProfileNumber) {
+    
+  }
+  createComboBoxForProfiles(adjustHeightDialog, onChangedSelection)
   
   height <- tclVar(10)
-  entry <- tkentry(h, width="10", textvariable = height)
-  tkgrid(tklabel(h, text="Profil Nummer: "),
-         comboMenu, entry, 
-         tklabel(h, text=" Meter"))
+  entry <-
+    tkentry(adjustHeightDialog,
+            width = '10',
+            textvariable = height)
+  tkgrid(entry, tklabel(adjustHeightDialog, text = ' Meter'))
   
   onOK <- function() {
-    tkdestroy(h)
-    tkfocus(tt)  
-    heightAdjustment(p[[as.integer(tclvalue(number))]], as.integer(tclvalue(height)))
+    tkdestroy(adjustHeightDialog)
+    tkfocus(topLevelWindow)
+    number <- as.integer(tclvalue(selectedProfileNumber))
+    profiles[[number]] <<-
+      adjustHeight(profiles[[number]], as.integer(tclvalue(height)))
   }
   
-  tkgrid(tkbutton(h, text="OK", command=onOK))
+  tkgrid(tkbutton(adjustHeightDialog, text = 'OK', command = onOK))
 }
 
 ###---GUI methods---####
 settings <- function() {
-  if (file.exists("config")) {
-    config <- read.table("config")
-    pointsize <- as.character(config[1,2])
+  if (file.exists('config')) {
+    config <- read.table('config')
+    pointsize <- as.character(config[1, 2])
+    colorRamp <- as.character(config[2, 2])
   } else {
     pointsize <- 10
+    colorRamp <- 'grey'
   }
   
-  s <- tktoplevel(tt)
-  tktitle(s) <- "Settings"
+  settingsDialog <- tktoplevel(topLevelWindow)
+  tktitle(settingsDialog) <- 'Settings'
   
-  SliderValue <- tclVar(pointsize)
-  slider <- tkscale(s, from=0, to=20,
-                    showvalue=T, variable=SliderValue,
-                    resolution=1, orient="horizontal")
-  tkgrid(tklabel(s, text="Pointsize: "), slider)
+  sliderValue <- tclVar(pointsize)
+  slider <- tkscale(
+    settingsDialog,
+    from = 0,
+    to = 20,
+    showvalue = T,
+    variable = sliderValue,
+    resolution = 1,
+    orient = 'horizontal'
+  )
+  tkgrid(tklabel(settingsDialog, text = 'Pointsize: '), slider)
   
-  colorType <- c("grey", "multicolored")
+  colorRamps <- c('grey', 'multicolored')
   
-  color <- tclVar("grey")
-  colorMenu <- ttkcombobox(s, values=colorType, textvariable=color, width=10)
-  tkgrid(tklabel(s, text="Color: "), colorMenu)
+  colorRamp <- tclVar(colorRamp)
+  colorComboBox <-
+    ttkcombobox(
+      settingsDialog,
+      values = colorRamps,
+      textvariable = colorRamp,
+      width = 10
+    )
+  tkgrid(tklabel(settingsDialog, text = 'Color: '), colorComboBox)
   
   onOK <- function() {
-    tkgrab.release(s)
-    tkdestroy(s)
-    tkfocus(tt)   
-    pointsize <<- tclvalue(SliderValue)
-    switch (tclvalue(color),
-            "grey" = {ramp <- c("black", "grey", "white")},
-            "multicolored" = {ramp <- c("blue", "green", "yellow", "orange", "red", "purple")})
-    colors <<- ramp
+    tkgrab.release(settingsDialog)
+    tkdestroy(settingsDialog)
+    tkfocus(topLevelWindow)
+    pointsize <- tclvalue(sliderValue)
+    assign("pointsize", pointsize, envir = .GlobalEnv)
+    
+    switch(tclvalue(colorRamp),
+           'grey' = {
+             colors <- c('black', 'grey', 'white')
+           },
+           'multicolored' = {
+             colors <- c('blue', 'green', 'yellow', 'orange', 'red', 'purple')
+           })
+    assign("colors", colors, envir = .GlobalEnv)
     
     # save settings to file
-    write.table(c("Pointsize" = pointsize, "Color" = as.vector(colors)), file="config", quote=FALSE, col.names=FALSE)
+    write.table(
+      c(
+        'Pointsize' = pointsize,
+        'ColorRamp' = tclvalue(colorRamp)
+      ),
+      file = 'config',
+      quote = FALSE,
+      col.names = FALSE
+    )
   }
   
-  tkgrid(tkbutton(s, text="OK", command=onOK))
+  tkgrid(tkbutton(settingsDialog, text = 'OK', command = onOK))
 }
 
 info <- function() {
-  tkmessageBox(title = "About this program",
-               message = "Developed by Anja Kleebaum \n
-Thanks to Prof. Dr. Klaus Bitzer & Richard Regner", 
-               icon = "info", type = "ok")
+  tkmessageBox(
+    title = 'About this program',
+    message = ' Developed by Anja Kleebaum\n Thanks to Prof. Dr. Klaus Bitzer & Richard Regner',
+    icon = 'info',
+    type = 'ok'
+  )
 }
 
-# Save Files
-saveFile <- function(type) { 
-  fileType = "{{GE3D Project (R Workspace)} {.RData}} {{All files} *}"  
-  multiFiles <- FALSE
+# Save File
+saveFile <- function(type) {
+  fileType = '{{Geoelectrics Project (R Workspace)} {.RData}} {{All files} *}'
   save <- function(fileName) {
-    save.image(file=fileName, ascii=TRUE)
+    save.image(file = fileName, ascii = TRUE)
   }
   
-  fileName <- as.character(tkgetSaveFile(filetypes = fileType))  
+  fileName <- as.character(tkgetSaveFile(filetypes = fileType))
   
   if (!nchar(fileName)) {
-    tkmessageBox(message = "No file was selected!")
+    tkmessageBox(message = 'No file was selected.')
   } else {
     save(fileName)
   }
 }
 
 # Open Files
-openFile <- function(type) {  
+openFile <- function(type) {
   switch(type,
-         "xyz" = {
-           fileType = "{{XYZ Files} {.xyz}} {{All files} *}"  
+         'processed' = {
+           fileType = '{{XYZ Files} {.xyz}} {{All files} *}'
            multiFiles <- TRUE
-           open <- function(fileName) {}
+           open <- function(fileName) {
+             
+           }
          },
-         "raw" = {
-           fileType = "{{Rawdata} {.dat}} {{Rawdata} {.txt}} {{All files} *}"  
+         'raw' = {
+           fileType = '{{Rawdata} {.dat}} {{Rawdata} {.txt}} {{All files} *}'
            multiFiles <- TRUE
-           open <- function(fileName) {}
+           open <- function(fileName) {
+             
+           }
          },
-         "rdata" = {
-           fileType = "{{GE3D Project (R Workspace)} {.RData}} {{All files} *}"  
+         'rdata' = {
+           fileType = '{{Geoelectrics Project (R Workspace)} {.RData}} {{All files} *}'
            multiFiles <- FALSE
            open <- function(fileName) {
              load(fileName, envir = .GlobalEnv)
-           }})
+           }
+         })
   
-  fileName <- as.character(tkgetOpenFile(filetypes = fileType, multiple=FALSE))  
+  fileName <-
+    as.character(tkgetOpenFile(filetypes = fileType, multiple = FALSE))
   
   if (!nchar(fileName)) {
-    tkmessageBox(message = "No file was selected!")
+    tkmessageBox(message = 'No file was selected!')
   } else {
     open(fileName)
     return(fileName)
-  } 
+  }
 }
 
-# Organize Project
-organizeProject <- function() {
-  or <- tktoplevel(tt)
-  tktitle(or) <- "Organize current GE3D Project"
-  tkwm.minsize(or, 700, 200)
-  comboMenu <- comboMenuProfiles(or)
-
-  tkbind(comboMenu, "<<ComboboxSelected>>", 
-         function() {
-           currentProfileNumber <- as.integer(tclvalue(number))           
-           try(tkdestroy(profileFrame))           
-           assign("profileFrame", tkframe(or, relief="groove", borderwidth=0), envir = .GlobalEnv)
-           tkgrid(profileFrame)
-           
-           xyzLabel <- tklabel(profileFrame, text="xyz-File:")
-           xyzAddress <- tclVar(p[[currentProfileNumber]]@xyzData@address)
-           xyzEntry <- tkentry(profileFrame, width="60", textvariable = xyzAddress, state = "readonly")
-           xyzButton <- tkbutton(profileFrame, text="Change", command = function() {
-             tclvalue(xyzAddress) <- openFile("xyz")
-             p[currentProfileNumber] <- setXyzAddress(p[[currentProfileNumber]], tclvalue(xyzAddress))
-             assign("p", p, envir = .GlobalEnv)
-           })
-           tkgrid(xyzLabel, xyzEntry, xyzButton)
-           
-           rawLabel <- tklabel(profileFrame, text="raw-data-File:")
-           rawAddress <- tclVar(p[[currentProfileNumber]]@rawData@address)
-           rawEntry <- tkentry(profileFrame, width="60", textvariable = rawAddress, state = "readonly")
-           rawButton <- tkbutton(profileFrame, text="Change", command = function() {
-             tclvalue(rawAddress) <- openFile("raw")
-             p[currentProfileNumber] <- setRawAddress(p[[currentProfileNumber]], tclvalue(rawAddress))
-             assign("p", p, envir = .GlobalEnv)
-           })
-           tkgrid(rawLabel, rawEntry, rawButton)
-           
-           gpsLabel <- tklabel(profileFrame, text="gps-File:")
-           gpsAddress <- tclVar(p[[currentProfileNumber]]@gpsCoordinates@address)
-           gpsEntry <- tkentry(profileFrame, width="60", textvariable = gpsAddress, state = "readonly")
-           gpsButton <- tkbutton(profileFrame, text="Change", command = function() {
-             tclvalue(gpsAddress) <- openFile("raw")
-             p[currentProfileNumber] <- setGpsAddress(p[[currentProfileNumber]], tclvalue(gpsAddress))
-             assign("p", p, envir = .GlobalEnv)
-           })
-           tkgrid(gpsLabel, gpsEntry, gpsButton)
-         })
+# Update Project
+updateProject <- function() {
+  updateProjectDialog <- tktoplevel(topLevelWindow)
+  tktitle(updateProjectDialog) <-
+    'Update geoelectrics project settings'
+  tkwm.minsize(updateProjectDialog, 700, 200)
   
   onOK <- function() {
-    tkgrab.release(or)
-    tkdestroy(or)
-    tkfocus(tt)
-    assign("p", p, envir = .GlobalEnv)
-  }  
+    tkgrab.release(updateProjectDialog)
+    tkdestroy(updateProjectDialog)
+    tkfocus(topLevelWindow)
+    assign('profiles', profiles, envir = .GlobalEnv)
+  }
   
-  ok.but <- tkbutton(or, text="OK", command = onOK)
-  tkgrid(comboMenu, ok.but, sticky="w")
+  onChangedSelection <- function(selectedProfileNumber) {
+    try(tkdestroy(profileFrame))
+    profileFrame <- tkframe(updateProjectDialog,
+                            relief = 'groove',
+                            borderwidth = 0)
+    assign('profileFrame', profileFrame, envir = .GlobalEnv)
+    tkgrid(profileFrame)
+    
+    processedDataLabel <-
+      tklabel(profileFrame, text = 'Processed data file:')
+    processedDataAddress <-
+      tclVar(p[[selectedProfileNumber]]@processedData@address)
+    processedDataEntry <-
+      tkentry(
+        profileFrame,
+        width = '60',
+        textvariable = processedDataAddress,
+        state = 'readonly'
+      )
+    processedDataButton <-
+      tkbutton(
+        profileFrame,
+        text = 'Change',
+        command = function() {
+          tclvalue(processedDataAddress) <- openFile('processed')
+          profiles[[selectedProfileNumber]]@processedData <-
+            new('ProcessedData', tclvalue(processedDataAddress))
+          assign('profiles', profiles, envir = .GlobalEnv)
+        }
+      )
+    tkgrid(processedDataLabel,
+           processedDataEntry,
+           processedDataButton)
+    
+    rawLabel <-
+      tklabel(profileFrame, text = 'Raw data file:')
+    rawAddress <-
+      tclVar(profiles[[selectedProfileNumber]]@rawData@address)
+    rawEntry <-
+      tkentry(
+        profileFrame,
+        width = '60',
+        textvariable = rawAddress,
+        state = 'readonly'
+      )
+    rawButton <-
+      tkbutton(
+        profileFrame,
+        text = 'Change',
+        command = function() {
+          tclvalue(rawAddress) <- openFile('raw')
+          profiles[[selectedProfileNumber]]@rawData <-
+            new('RawData', tclvalue(rawAddress))
+          assign('profiles', profiles, envir = .GlobalEnv)
+        }
+      )
+    tkgrid(rawLabel, rawEntry, rawButton)
+    
+    gpsLabel <- tklabel(profileFrame, text = 'GPS file:')
+    gpsAddress <-
+      tclVar(profiles[[selectedProfileNumber]]@gpsCoordinates@address)
+    gpsEntry <-
+      tkentry(
+        profileFrame,
+        width = '60',
+        textvariable = gpsAddress,
+        state = 'readonly'
+      )
+    gpsButton <-
+      tkbutton(
+        profileFrame,
+        text = 'Change',
+        command = function() {
+          tclvalue(gpsAddress) <- openFile('raw')
+          profiles[[selectedProfileNumber]]@gpsCoordinates <-
+            new('GpsCoordinates', tclvalue(gpsAddress))
+          assign('profiles', profiles, envir = .GlobalEnv)
+        }
+      )
+    tkgrid(gpsLabel, gpsEntry, gpsButton)
+    
+    okButton <- tkbutton(profileFrame, text = 'OK', command = onOK)
+    tkgrid(
+      okButton,
+      sticky = 'e',
+      columnspan = 3,
+      pady = c(10, 10)
+    )
+  }
+  
+  createComboBoxForProfiles(updateProjectDialog, onChangedSelection)
   
   return(p)
 }
 
-modalDialog <- function(title, question, entryInit, entryWidth = 20,
-                        returnValOnCancel = "ID_CANCEL") {
-  dlg <- tktoplevel()
-  tkwm.deiconify(dlg)
-  tkgrab.set(dlg)
-  tkfocus(dlg)
-  tkwm.title(dlg, title)
+modalDialog <- function(title,
+                        question,
+                        entryInit,
+                        entryWidth = 20,
+                        returnValOnCancel = 'ID_CANCEL') {
+  dialog <- tktoplevel()
+  tkwm.deiconify(dialog)
+  tkgrab.set(dialog)
+  tkfocus(dialog)
+  tkwm.title(dialog, title)
   textEntryVarTcl <- tclVar(paste(entryInit))
-  textEntryWidget <- tkentry(dlg, width = paste(entryWidth),
+  textEntryWidget <- tkentry(dialog,
+                             width = paste(entryWidth),
                              textvariable = textEntryVarTcl)
-  tkgrid(tklabel(dlg, text = "       "))
-  tkgrid(tklabel(dlg, text = question), textEntryWidget)
-  tkgrid(tklabel(dlg, text = "       "))
-  ReturnVal <- returnValOnCancel
+  tkgrid(tklabel(dialog, text = '       '))
+  tkgrid(tklabel(dialog, text = question), textEntryWidget)
+  tkgrid(tklabel(dialog, text = '       '))
+  returnVal <- returnValOnCancel
   
   onOK <- function() {
-    ReturnVal <<- tclvalue(textEntryVarTcl)
-    tkgrab.release(dlg)
-    tkdestroy(dlg)
-    tkfocus(tt)
+    returnVal <<- tclvalue(textEntryVarTcl)
+    tkgrab.release(dialog)
+    tkdestroy(dialog)
+    tkfocus(topLevelWindow)
   }
   onCancel <- function() {
-    ReturnVal <<- returnValOnCancel
-    tkgrab.release(dlg)
-    tkdestroy(dlg)
-    tkfocus(tt)
+    tkgrab.release(dialog)
+    tkdestroy(dialog)
+    tkfocus(topLevelWindow)
+    return(returnValOnCancel)
   }
-  OK.but <- tkbutton(dlg, text = "   OK   ", command = onOK)
-  Cancel.but <- tkbutton(dlg, text = " Cancel ", command = onCancel)
-  tkgrid(OK.but, Cancel.but)
-  tkgrid(tklabel(dlg, text = "    "))
+  okButton <- tkbutton(dialog, text = '   OK   ', command = onOK)
+  cancelButton <-
+    tkbutton(dialog, text = ' Cancel ', command = onCancel)
+  tkgrid(okButton, cancelButton)
+  tkgrid(tklabel(dialog, text = '    '))
   
-  tkfocus(dlg)
-  tkbind(dlg, "<Destroy>", function() {tkgrab.release(dlg); tkfocus(tt)})
-  tkbind(textEntryWidget, "<Return>", onOK)
-  tkwait.window(dlg)
+  tkfocus(dialog)
+  tkbind(dialog, '<Destroy>', function() {
+    tkgrab.release(dialog)
+    tkfocus(topLevelWindow)
+  })
+  tkbind(textEntryWidget, '<Return>', onOK)
+  tkwait.window(dialog)
   
-  return(ReturnVal)
+  return(returnVal)
 }
 
 dropCurrentProject <- function() {
-  ReturnVal <- tkmessageBox(message = "Do you want to save before dropping current project?",
-                            icon = "question", type = "yesnocancel", default = "yes")
-  if(tclvalue(ReturnVal) == "cancel") {
+  returnVal <-
+    tkmessageBox(
+      message = 'Do you want to save before dropping current project?',
+      icon = 'question',
+      type = 'yesnocancel',
+      default = 'yes'
+    )
+  if (tclvalue(returnVal) == 'cancel') {
     return()
   }
-  else if(tclvalue(ReturnVal) == "yes") {
-    saveFile("rdata")
+  else if (tclvalue(returnVal) == 'yes') {
+    saveFile('rdata')
   }
 }
 
 newProject <- function() {
-  ReturnVal <- tkmessageBox(message = "Do you want to save before dropping current project?",
-                            icon = "question", type = "yesnocancel", default = "yes")
-  if(tclvalue(ReturnVal) == "cancel") {
+  returnVal <-
+    tkmessageBox(
+      message = 'Do you want to save before dropping current project?',
+      icon = 'question',
+      type = 'yesnocancel',
+      default = 'yes'
+    )
+  if (tclvalue(returnVal) == 'cancel') {
     return()
   }
-  else if(tclvalue(ReturnVal) == "yes") {
-    saveFile("rdata")
+  else if (tclvalue(returnVal) == 'yes') {
+    saveFile('rdata')
   }
   
-  try(rm(p, envir = .GlobalEnv))
+  try(rm(profiles, envir = .GlobalEnv))
   
-  ReturnVal <- modalDialog("Profile Number", "Insert Profile Number", 1, 5)
+  numberOfProfiles <-
+    modalDialog('Create a new geoelctrics project',
+                'Insert Number of Profiles',
+                1,
+                5)
+  
   profiles <- list()
-  for(i in 1:as.numeric(ReturnVal)) {
-    profiles[i] <- new("Profile",
+  for (i in 1:as.numeric(numberOfProfiles)) {
+    profiles[i] <- new('Profile',
                        number = i)
   }
-  return(profiles)
+  assign('profiles', profiles, envir = .GlobalEnv)
+  updateProject()
 }
 
 ### menubar
-topMenu <- tkmenu(tt)
-tkconfigure(tt, menu = topMenu)
+topMenu <- tkmenu(topLevelWindow)
+tkconfigure(topLevelWindow, menu = topMenu)
 fileMenu <- tkmenu(topMenu, tearoff = FALSE)
 editMenu <- tkmenu(topMenu, tearoff = FALSE)
 plot2dMenu <- tkmenu(topMenu, tearoff = FALSE)
@@ -323,95 +484,209 @@ helpMenu <- tkmenu(topMenu, tearoff = FALSE)
 
 # cascaded menus
 openFileMenu <- tkmenu(topMenu, tearoff = FALSE)
-saveFileMenu <- tkmenu(topMenu, tearoff = FALSE) 
-plot2dRaw <- tkmenu(topMenu, tearoff = FALSE)
-plot2dXyz <- tkmenu(topMenu, tearoff = FALSE)
-plot3dSingle <- tkmenu(topMenu, tearoff = FALSE) 
-plot3dFew <- tkmenu(topMenu, tearoff = FALSE)
-plot3dAll <- tkmenu(topMenu, tearoff = FALSE) 
+saveFileMenu <- tkmenu(topMenu, tearoff = FALSE)
+plotRawDataMenu <- tkmenu(topMenu, tearoff = FALSE)
+plotProcessedDataMenu <- tkmenu(topMenu, tearoff = FALSE)
+plot3dSingle <- tkmenu(topMenu, tearoff = FALSE)
+plot3dAll <- tkmenu(topMenu, tearoff = FALSE)
 example <- tkmenu(topMenu, tearoff = FALSE)
 
-tkadd(topMenu, "cascade", label = "File", menu = fileMenu)
-tkadd(topMenu, "cascade", label = "Edit", menu = editMenu)
-tkadd(topMenu, "cascade", label = "Plot 2D", menu = plot2dMenu)
-tkadd(topMenu, "cascade", label = "Plot 3D", menu = plot3dMenu)
-tkadd(topMenu, "cascade", label = "Help", menu = helpMenu)
+tkadd(topMenu, 'cascade', label = 'File', menu = fileMenu)
+tkadd(topMenu, 'cascade', label = 'Edit', menu = editMenu)
+tkadd(topMenu, 'cascade', label = 'Plot 2D', menu = plot2dMenu)
+tkadd(topMenu, 'cascade', label = 'Plot 3D', menu = plot3dMenu)
+tkadd(topMenu, 'cascade', label = 'Help', menu = helpMenu)
 
-tkadd(fileMenu, "command", label = "New Project", command = function() {
-  assign("p", newProject(), envir = .GlobalEnv)
-  organizeProject()})
-tkadd(fileMenu, "command", label = "Open Project", command = function() openFile("rdata"))
-tkadd(fileMenu, "command", label = "Save Project", command = function() saveFile("rdata"))
-tkadd(fileMenu, "command", label = "Organize Project", command = function() organizeProject())
-tkadd(fileMenu, "separator")
-tkadd(fileMenu, "command", label = "Quit Program", command = function() {
-  ReturnVal <- tkmessageBox(message = "Do you want to save before quitting?",
-               icon = "question", type = "yesnocancel", default = "yes")
-  if(tclvalue(ReturnVal) == "cancel") {
-    return()
+tkadd(
+  fileMenu,
+  'command',
+  label = 'New Project',
+  command = function() {
+    newProject()
   }
-  else if(tclvalue(ReturnVal) == "yes") {
-    saveFile("rdata")
+)
+tkadd(
+  fileMenu,
+  'command',
+  label = 'Open Project',
+  command = function()
+    openFile('rdata')
+)
+tkadd(
+  fileMenu,
+  'command',
+  label = 'Save Project',
+  command = function()
+    saveFile('rdata')
+)
+tkadd(
+  fileMenu,
+  'command',
+  label = 'Update Project',
+  command = function()
+    updateProject()
+)
+tkadd(fileMenu, 'separator')
+tkadd(
+  fileMenu,
+  'command',
+  label = 'Quit Program',
+  command = function() {
+    ReturnVal <-
+      tkmessageBox(
+        message = 'Do you want to save before quitting?',
+        icon = 'question',
+        type = 'yesnocancel',
+        default = 'yes'
+      )
+    if (tclvalue(ReturnVal) == 'cancel') {
+      return()
+    }
+    else if (tclvalue(ReturnVal) == 'yes') {
+      saveFile('rdata')
+    }
+    tkdestroy(topLevelWindow)
   }
-  tkdestroy(tt)
-  })
+)
 
-tkadd(editMenu, "command", label = "Settings", command = function() settings())
-tkadd(editMenu, "separator")
-tkadd(editMenu, "command", label = "Height Adjustment", command = function() {
-  heightAdjustmentGui()
-})
+tkadd(
+  editMenu,
+  'command',
+  label = 'Plotting Settings',
+  command = function()
+    settings()
+)
+tkadd(editMenu, 'separator')
+tkadd(
+  editMenu,
+  'command',
+  label = 'Height Adjustment',
+  command = function() {
+    adjustHeightGui()
+  }
+)
 
-tkadd(plot2dMenu, "cascade", label = "Rawdata", menu = plot2dRaw)
-tkadd(plot2dMenu, "cascade", label = "After Inversion", menu = plot2dXyz)
+tkadd(plot2dMenu, 'cascade', label = 'Rawdata', menu = plotRawDataMenu)
+tkadd(plot2dMenu, 'cascade', label = 'After Inversion', menu = plotProcessedDataMenu)
 
-tkadd(plot3dMenu, "command", label = "Single Profiles", command = function() choosePlottingProfile(plot3dXyz))
-tkadd(plot3dMenu, "command", label = "All Profiles", command = function() { 
-  for(x in p) {    
-    plot3dXyz(p[[x@number]])
-  }})
+tkadd(
+  plot3dMenu,
+  'command',
+  label = 'Single Profiles',
+  command = function()
+    plotProfileWithPlottingFunction(function(profile) {
+      plot3d(profile, psize = pointsize, col = colors)
+    })
+)
+tkadd(
+  plot3dMenu,
+  'command',
+  label = 'All Profiles',
+  command = function() {
+    for (i in 1:length(profiles)) {
+      plot3d(profiles[[i]], psize = pointsize, col = colors)
+    }
+  }
+)
 
-tkadd(helpMenu, "cascade", label = "Load Example", menu = example)
-tkadd(helpMenu, "separator")
-tkadd(helpMenu, "command", label = "About Program", command = info)
+tkadd(helpMenu, 'cascade', label = 'Load Example', menu = example)
+tkadd(helpMenu, 'separator')
+tkadd(helpMenu, 'command', label = 'About Program', command = info)
 
-tkadd(plot2dRaw, "command", label = "Points", command = function() choosePlottingProfile(plotRawGui))
-tkadd(plot2dRaw, "command", label = "Levels", command = function() choosePlottingProfile(levelplotRawGui))
+plotProfileWithPlottingFunction <- function(plottingFunction) {
+  resetFrames()
+  onChangedSelection <- function(selectedProfileNumber) {
+    plottingFunction(profiles[[selectedProfileNumber]])
+  }
+  createComboBoxForProfiles(topLevelWindow, onChangedSelection)
+}
 
-tkadd(plot2dXyz, "command", label = "Points", command = function() choosePlottingProfile(plotXyzGui))
-tkadd(plot2dXyz, "command", label = "Levels", command = function() choosePlottingProfile(levelplotXyzGui))
-tkadd(plot2dXyz, "command", label = "Points with Topography", command = function() choosePlottingProfile(plotXyzHeightGui))
-tkadd(plot2dXyz, "command", label = "Levels with Topography", command = function() choosePlottingProfile(levelplotXyzHeightGui))
+tkadd(
+  plotRawDataMenu,
+  'command',
+  label = 'Points',
+  command = function()
+    plotProfileWithPlottingFunction(plotRawDataGui)
+)
+tkadd(
+  plotRawDataMenu,
+  'command',
+  label = 'Levels',
+  command = function()
+    plotProfileWithPlottingFunction(levelplotRawDataGui)
+)
 
-tkadd(example, "command", label = "Sinkhole", command = function() {data(sinkhole); p<<-list(sinkhole@profiles[[1]],sinkhole@profiles[[2]],sinkhole@profiles[[3]])})
+tkadd(
+  plotProcessedDataMenu,
+  'command',
+  label = 'Points',
+  command = function()
+    plotProfileWithPlottingFunction(plotProcessedDataGui)
+)
+tkadd(
+  plotProcessedDataMenu,
+  'command',
+  label = 'Levels',
+  command = function()
+    plotProfileWithPlottingFunction(levelplotProcessedDataGui)
+)
+tkadd(
+  plotProcessedDataMenu,
+  'command',
+  label = 'Points with Topography',
+  command = function()
+    plotProfileWithPlottingFunction(plotProcessedDataWithTopoGui)
+)
+tkadd(
+  plotProcessedDataMenu,
+  'command',
+  label = 'Levels with Topography',
+  command = function()
+    plotProfileWithPlottingFunction(levelplotProcessedDataWithTopoGui)
+)
+
+tkadd(
+  example,
+  'command',
+  label = 'Sinkhole',
+  command = function() {
+    data(sinkhole)
+    profiles <-
+      list(sinkhole@profiles[[1]], sinkhole@profiles[[2]], sinkhole@profiles[[3]])
+    assign('profiles', profiles, envir = .GlobalEnv)
+  }
+)
 
 ### frames
+resetFrames <- function() {
+  deleteFrame(frameLeft)
+  deleteFrame(frameRight)
+  frameLeft <<- newFrame('frameLeft')
+  frameRight <<- newFrame('frameRight')
+  return(tkgrid(frameLeft, frameRight))
+}
+
 deleteFrame <- function(frameName) {
   tkdestroy(frameName)
 }
 
 newFrame <- function(frameName) {
-  frameName <<- tkframe(tt, relief="groove", borderwidth=0)
-}
-
-resetFrames <- function() {
-  deleteFrame(frameLeft)
-  deleteFrame(frameRight)
-  frameLeft <<- newFrame("frameLeft")
-  frameRight <<- newFrame("frameRight")
-  return(tkgrid(frameLeft, frameRight))
+  frameName <<-
+    tkframe(topLevelWindow, relief = 'groove', borderwidth = 0)
 }
 
 # left frame
-frameLeft <<- newFrame("frameLeft")
-tkgrid(tklabel(frameLeft, text=""))
+frameLeft <- newFrame('frameLeft')
+tkgrid(tklabel(frameLeft, text = ''))
 
 # right frame
-frameRight <<- newFrame("frameRight")
-tkgrid(tklabel(frameRight, text="Welcome to the Software for 3D-Presentation of 2D-Geoelectric Profiles"))
+frameRight <- newFrame('frameRight')
+tkgrid(
+  tklabel(frameRight, text = 'Welcome to the Software for 3D-Presentation of 2D-Geoelectric Profiles')
+)
 
 # create frames
 tkgrid(frameLeft, frameRight)
 
-# don't terminate 
-Sys.sleep(360000)
+# don't terminate
+#Sys.sleep(360000)